teaching_scripts/requirements.txt
google-api-python-client
imagehash
pillow
pillow_heif
py-applescript
